// src/app/real-time/models/mensaje.model.ts
export class Mensaje {

    id: string;
    uniqueId: string;
    contextId: string;
    parentId: string;
    content: string;
    source: string;
    sourceAlias: string;
    sourceEmail: string;
    sourcePhone: string;
    target: string;
    targetAlias: string;
    targetEmail: string;
    targetPhone: string;
    value: string;
    dataSource: string;
    dataSourceId: string;
    dataSourceType: string;
    dataSourceAlias: string;
    createdBy: string;
    createdAt: string;
    updatedBy: string;
    updatedAt: string;
    payload: string;

}